# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/seaport": {
            "get": {
                "tags": [
                    "seaport"
                ],
                "description": "Return a list seaport",
                "operationId": "getListSeaport",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number"
                    },
                    {
                        "name": "pagesize",
                        "in": "query",
                        "description": "number of records in page"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "seaport"
                ],
                "description": "Create a seaport",
                "operationId": "createSeaport",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name of seaport",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "code of seaport",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "max_draft",
                        "in": "query",
                        "description": "max_draft of seaport",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/seaport/{id}": {
            "put": {
                "tags": [
                    "seaport"
                ],
                "description": "Update a seaport",
                "operationId": "updateSeaport",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name need update",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "code need update",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "max_draft",
                        "in": "query",
                        "description": "max_draft need update",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of seaport to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "seaport"
                ],
                "description": "Delete a seaport",
                "operationId": "deleteSeaport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of seaport to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    }
}